Updates to build on windows
