+++ client_windows.go
@@ -1,0 +1,30 @@
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package main
+
+import (
+	"time"
+
+	"github.com/Microsoft/go-winio"
+)
+
+func isSocketAccessible(s string) error {
+	// test if we can access the pipe
+	timeout := 2 * time.Second
+	_, err := winio.DialPipe(s, &timeout)
+	return err
+}
