+++ go.mod
@@ -3,6 +3,7 @@ module github.com/containerd/nerdctl
 go 1.16
 
 require (
+	github.com/Microsoft/go-winio v0.4.17
 	github.com/compose-spec/compose-go v0.0.0-20210420125800-01e9e6b4c64c
 	github.com/containerd/cgroups v1.0.0
 	github.com/containerd/console v1.0.2
