+++ login_windows.go
@@ -1,0 +1,40 @@
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package main
+
+import (
+	"fmt"
+	"syscall"
+
+	"github.com/pkg/errors"
+	"golang.org/x/term"
+)
+
+func readPassword() (string, error) {
+	var fd int
+	if term.IsTerminal(int(syscall.Stdin)) {
+		fd = int(syscall.Stdin)
+	} else {
+		return "", fmt.Errorf("error allocating terminal")
+	}
+	bytePassword, err := term.ReadPassword(fd)
+	if err != nil {
+		return "", errors.Wrap(err, "error reading password")
+	}
+
+	return string(bytePassword), nil
+}
