+++ run.go
@@ -32,7 +32,6 @@ import (
 	"github.com/containerd/containerd/cmd/ctr/commands/tasks"
 	"github.com/containerd/containerd/containers"
 	"github.com/containerd/containerd/oci"
-	"github.com/containerd/containerd/pkg/cap"
 	"github.com/containerd/containerd/plugin"
 	"github.com/containerd/containerd/runtime/restart"
 	gocni "github.com/containerd/go-cni"
@@ -273,7 +272,7 @@ func runAction(clicontext *cli.Context) error {
 		oci.WithMounts([]specs.Mount{
 			{Type: "cgroup", Source: "cgroup", Destination: "/sys/fs/cgroup", Options: []string{"ro", "nosuid", "noexec", "nodev"}},
 		}),
-		oci.WithoutRunMount, // unmount default tmpfs on "/run": https://github.com/containerd/nerdctl/issues/157
+		WithoutRunMount(), // unmount default tmpfs on "/run": https://github.com/containerd/nerdctl/issues/157
 	)
 
 	rootfsOpts, rootfsCOpts, ensuredImage, err := generateRootfsOpts(ctx, client, clicontext, id)
@@ -734,53 +733,4 @@ func propagateContainerdLabelsToOCIAnnotations() oci.SpecOpts {
 	return func(ctx context.Context, oc oci.Client, c *containers.Container, s *oci.Spec) error {
 		return oci.WithAnnotations(c.Labels)(ctx, oc, c, s)
 	}
-}
-
-func runBashComplete(clicontext *cli.Context) {
-	coco := parseCompletionContext(clicontext)
-	if coco.boring {
-		defaultBashComplete(clicontext)
-		return
-	}
-	if coco.flagTakesValue {
-		w := clicontext.App.Writer
-		switch coco.flagName {
-		case "restart":
-			fmt.Fprintln(w, "always")
-			fmt.Fprintln(w, "no")
-			return
-		case "pull":
-			fmt.Fprintln(w, "always")
-			fmt.Fprintln(w, "missing")
-			fmt.Fprintln(w, "never")
-			return
-		case "cgroupns":
-			fmt.Fprintln(w, "host")
-			fmt.Fprintln(w, "private")
-			return
-		case "security-opt":
-			fmt.Fprintln(w, "seccomp=")
-			fmt.Fprintln(w, "apparmor="+defaults.AppArmorProfileName)
-			fmt.Fprintln(w, "no-new-privileges")
-			return
-		case "cap-add", "cap-drop":
-			for _, c := range cap.Known() {
-				// "CAP_SYS_ADMIN" -> "sys_admin"
-				s := strings.ToLower(strings.TrimPrefix(c, "CAP_"))
-				fmt.Fprintln(w, s)
-			}
-			return
-		case "net", "network":
-			bashCompleteNetworkNames(clicontext, nil)
-			return
-		}
-		defaultBashComplete(clicontext)
-		return
-	}
-	// show image names, unless we have "--rootfs" flag
-	if clicontext.Bool("rootfs") {
-		defaultBashComplete(clicontext)
-		return
-	}
-	bashCompleteImageNames(clicontext)
 }
